pyproject.toml
[tool.poetry]
name = "hentai-tgbot"
version = "0.1.0"
description = ""
authors = ["waydk <waydk.note@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
aiogram = "^3.1.1"
loguru = "^0.7.2"
hmtai = "^3.2.31"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
